Feature/dev chaos

+++ src/main/java/org/container/platform/common/api/chaos/ResourceUsageItem.java
@@ -16,8 +16,8 @@ import java.util.List;
 public class ResourceUsageItem {
     private List<String> time;
     private List<String> resourceName;
-    private List<List<Long>> cpu;
-    private List<List<Long>> memory;
+    private List<List<Integer>> cpu;
+    private List<List<Integer>> memory;
     private List<List<Integer>> appStatus;
 
     public ResourceUsageItem() {
